Refactor/notification4

+++ src/app/admin/mail/[filter]/[id]/MailFlow.tsx
@@ -8,7 +8,8 @@ import {
     destroyMailingListGroupRelationAction,
     destroyMailingListUserRelationAction
 } from '@/services/mail/actions'
-import { useSession } from '@/auth/session/useSession'
+import { mailAuth } from '@/services/mail/auth'
+import useAuthorizer from '@/hooks/useAuthorizer'
 import type { ActionReturn } from '@/services/actionTypes'
 import type { MailFlowObject, MailListTypes } from '@/services/mail/types'
 
@@ -29,68 +30,44 @@ export default function MailFlow({
     let userDestroy: DestroyFunction = null
     let addressExternalDestroy: DestroyFunction = null
 
-    const session = useSession()
-    const permissions = !session.loading ? session.session.permissions : []
+    const canAdminMailingList = useAuthorizer({
+        authorizer: mailAuth.destroyAliasMailingListRelation.dynamicFields({})
+    }).authorized
 
-    if (filter === 'mailingList') {
-        if (permissions.includes('MAILINGLIST_ALIAS_DESTROY')) {
-            aliasDestroy = async (mailAliasId: number) => {
-                const ret = await destroyAliasMailingListRelationAction({
-                    mailingListId: id,
-                    mailAliasId,
-                })
-                return {
-                    ...ret,
-                    data: null,
-                }
-            }
+    if (filter === 'mailingList' && canAdminMailingList) {
+        aliasDestroy = async (mailAliasId: number) => {
+            const ret = await destroyAliasMailingListRelationAction({
+                data: { mailingListId: id, mailAliasId },
+            })
+            return { ...ret, data: null }
         }
 
-        if (permissions.includes('MAILINGLIST_EXTERNAL_ADDRESS_DESTROY')) {
-            addressExternalDestroy = async (mailAddressExternalId: number) => {
-                const ret = await destroyMailingListExternalRelationAction({
-                    mailingListId: id,
-                    mailAddressExternalId,
-                })
-                return {
-                    ...ret,
-                    data: null,
-                }
-            }
+        addressExternalDestroy = async (mailAddressExternalId: number) => {
+            const ret = await destroyMailingListExternalRelationAction({
+                data: { mailingListId: id, mailAddressExternalId },
+            })
+            return { ...ret, data: null }
         }
 
-        if (permissions.includes('MAILINGLIST_USER_DESTROY')) {
-            userDestroy = async (userId: number) => {
-                const ret = await destroyMailingListUserRelationAction({
-                    mailingListId: id,
-                    userId,
-                })
-                return {
-                    ...ret,
-                    data: null,
-                }
-            }
+        userDestroy = async (userId: number) => {
+            const ret = await destroyMailingListUserRelationAction({
+                data: { mailingListId: id, userId },
+            })
+            return { ...ret, data: null }
         }
 
-        if (permissions.includes('MAILINGLIST_GROUP_DESTROY')) {
-            groupDestroy = async (groupId: number) => {
-                const ret = await destroyMailingListGroupRelationAction({
-                    mailingListId: id,
-                    groupId,
-                })
-                return {
-                    ...ret,
-                    data: null,
-                }
-            }
+        groupDestroy = async (groupId: number) => {
+            const ret = await destroyMailingListGroupRelationAction({
+                data: { mailingListId: id, groupId },
+            })
+            return { ...ret, data: null }
         }
     }
 
-    if (filter === 'alias' && permissions.includes('MAILINGLIST_ALIAS_DESTROY')) {
+    if (filter === 'alias' && canAdminMailingList) {
         mailingListDestroy = async (mailingListId: number) => {
             const ret = await destroyAliasMailingListRelationAction({
-                mailAliasId: id,
-                mailingListId,
+                data: { mailAliasId: id, mailingListId },
             })
             return {
                 ...ret,
@@ -99,11 +76,10 @@ export default function MailFlow({
         }
     }
 
-    if (filter === 'mailaddressExternal' && permissions.includes('MAILINGLIST_EXTERNAL_ADDRESS_DESTROY')) {
+    if (filter === 'mailaddressExternal' && canAdminMailingList) {
         mailingListDestroy = async (mailingListId: number) => {
             const ret = await destroyMailingListExternalRelationAction({
-                mailAddressExternalId: id,
-                mailingListId,
+                data: { mailAddressExternalId: id, mailingListId },
             })
             return {
                 ...ret,
@@ -112,11 +88,10 @@ export default function MailFlow({
         }
     }
 
-    if (filter === 'user' && permissions.includes('MAILINGLIST_USER_DESTROY')) {
+    if (filter === 'user' && canAdminMailingList) {
         mailingListDestroy = async (mailingListId: number) => {
             const ret = await destroyMailingListUserRelationAction({
-                userId: id,
-                mailingListId,
+                data: { userId: id, mailingListId },
             })
             return {
                 ...ret,
@@ -125,11 +100,10 @@ export default function MailFlow({
         }
     }
 
-    if (filter === 'group' && permissions.includes('MAILINGLIST_GROUP_DESTROY')) {
+    if (filter === 'group' && canAdminMailingList) {
         mailingListDestroy = async (mailingListId: number) => {
             const ret = await destroyMailingListGroupRelationAction({
-                groupId: id,
-                mailingListId,
+                data: { groupId: id, mailingListId },
             })
             return {
                 ...ret,
